Forcibly overwrite artifacts, modifying flags and using `sudo` if needed
- This commit implements [the proposed behavior for `install --force`](https://github.com/caskroom/homebrew-cask/pull/13966#issuecomment-220830387) when a target already exists and has either permission problems or is not owned by the user.

- The changes apply only when the `force` option is given.

- Reused the existing safeguard from the `.pkg` artifact to prevent deleting important directories by bug or mistake

- The two existing blacklists `SYSTEM_DIRS` and `UNDELETABLE_DIRS` have been consolidated into the `Hbc::MacOS` module.

- `UNDELETABLE_DIRS` now also contains all the entries from `SYSTEM_DIRS` which was a to-do anyway.

- The two blacklists are now also frozen for good measure.

- The utility method `permissions_rmtree` was moved to `Hbc::Utils`.

- The `tried_permissions` part in `Utils` now falls back correctly when there are also ownership issues at the same time.

- Introduced a separate `current_user` method for mocking.

- Added an optional feature to `FakeSystemCommand` so it can now act as a proxy to `SystemCommand`.

- Added tests for various `permissions_rmtree` cases.

+++ lib/hbc/artifact/moved.rb
@@ -22,7 +22,7 @@ class Hbc::Artifact::Moved < Hbc::Artifact::Base
     each_artifact do |artifact|
       load_specification(artifact)
       next unless preflight_checks
-      delete if File.exist?(target)
+      delete if File.exist?(target) && force
       move
     end
   end
@@ -91,7 +91,15 @@ class Hbc::Artifact::Moved < Hbc::Artifact::Base
 
   def delete
     ohai "Removing #{english_name}: '#{target}'"
-    target.rmtree
+    case
+    when Hbc::MacOS.undeletable?(target)
+      raise Hbc::CaskError.new(
+        "Cannot remove undeletable #{english_name}")
+    when force
+      Hbc::Utils.permissions_rmtree(target, command: @command)
+    else
+      target.rmtree
+    end
   end
 
   def summarize_artifact(artifact_spec)
